backend/requirements.txt
faiss-cpu
openai
tiktoken
docx2txt
pillow
pypdf
tenacity
botbuilder-integration-aiohttp>=4.14.4
python-dotenv
markdownify
langchain==0.0.347
langchain-experimental==0.0.44
openai==1.3.7
python-dotenv
# azure-ai-formrecognizer
azure-storage-blob
rarfile